LFS-34: User Management Page Put dialogs into individual jsx files. Split Userboard component into UsersManager and GroupsManager components for users and groups.
