+++ modules/homepage/src/main/frontend/src/themePage/Userboard/userboard.jsx
@@ -17,195 +17,13 @@
 
 import React from "react";
 
-import PropTypes from "prop-types";
+import UsersManager from "./usersmanager.jsx";
+import GroupsManager from "./groupsmanager.jsx";
 
-import  {withStyles} from "@material-ui/core/styles";
-
-import {Button, Table, TableBody, TableHead, TableRow, TableCell, TableFooter, TablePagination, Checkbox, Hidden, Dialog, DialogTitle, DialogActions, DialogContent, TextField} from "@material-ui/core"
-
-import GridItem from "material-dashboard-react/dist/components/Grid/GridItem.js";
-import GridContainer from "material-dashboard-react/dist/components/Grid/GridContainer.js";
-import Card from "material-dashboard-react/dist/components/Card/Card.js";
-import CardHeader from "material-dashboard-react/dist/components/Card/CardHeader.js";
-import CardBody from "material-dashboard-react/dist/components/Card/CardBody.js";
-import CardFooter from "material-dashboard-react/dist/components/Card/CardFooter"
-//import { Avatar } from "@material-ui/core";
-
-import userboardStyle from './userboardStyle.jsx';
-
-import {CreateUserDialogue, DeleteUserDialogue, ChangeUserPasswordDialogue, CreateGroupDialogue, DeleteGroupDialogue, AddUserToGroupDialogue, RemoveUserFromGroupDialogue} from './userboardDialogues.jsx';
 
 class Userboard extends React.Component {
   constructor(props) {
     super(props);
-    this.state = {
-      users: [],
-      groups: [],
-      userFilter: null,
-      groupFilter: null,
-
-      currentUserName: "",
-      currentUserIndex: -1,
-      returnedUserRows: 0,
-      totalUserRows: 0,
-
-      currentGroupName: "",
-      currentGroupIndex: -1,
-      returnedGroupRows: 0,
-      totalGroupRows: 0,
-
-      deployCreateUser: false,
-      deployDeleteUser: false,
-      deployChangeUserPassword: false,
-      deployCreateGroup: false,
-      deployDeleteGroup: false,
-      deployAddGroupUsers: false,
-      deployRemoveGroupUsers: false,
-      deployMobileUserDialog: false,
-      deployMobileGroupDialog: false
-    };
-
-    //this.userColumnNames = [{id: "name", label: "User Names"}];
-    //this.groupColumnNames = [{id: "name", label: "Group Names"}];
-  }
-
-  clearSelectedUser () {
-    this.setState(
-      {
-        currentUserName: "",
-        currentUserIndex: -1,
-        returnedUserRows: 0,
-        totalUserRows: 0
-      }
-    );
-  }
-
-  clearSelectedGroup () {
-    this.setState(
-      {
-        currentGroupName: "",
-        currentGroupIndex: -1,
-        returnedGroupRows: 0,
-        totalGroupRows: 0
-      }
-    );
-  }
-
-  handleLoadUsers (filter, offset, limit) {
-    let url = new URL("http://localhost:8080/home/users.json");
-
-    if (filter !== null && filter !== "") {
-      url.searchParams.append('filter', filter);
-    }
-
-    if (offset !== null) {
-      url.searchParams.append('offset', offset);
-    }
-
-    if (limit !== null) {
-      url.searchParams.append('limit', limit);
-    }
-
-    fetch(url,
-      {
-        method: 'GET',
-        headers: {
-          'Authorization': 'Basic ' + btoa('admin:admin')
-        },
-    })
-    .then((response) => {
-      return response.json();
-    })
-    .then((data) => {
-      this.clearSelectedUser();
-      this.setState(
-        {
-          returnedUserRows: data.returnedrows, 
-          totalUserRows: data.totalrows,
-          users: data.rows
-        }
-      );
-    })
-    .catch((error) => {
-      console.log(error);
-    });
-  }
-
-  handleLoadGroups (filter, offset, limit) {
-    let url = new URL ("http://localhost:8080/home/groups.json");
-
-    if (filter !== null && filter !== "") {
-      url.searchParams.append('filter', filter);
-    }
-
-    if (offset !== null) {
-      url.searchParams.append('offset', offset);
-    }
-
-    if (limit !== null) {
-      url.searchParams.append('limit', limit);
-    }
-
-    fetch(url,
-      {
-        method: 'GET',
-        headers: {
-          'Authorization' : 'Basic' + btoa('admin:admin')
-        }
-    })
-    .then((response) => {
-      return response.json();
-    })
-    .then((data) => {
-      this.clearSelectedGroup();
-      this.setState(
-        {
-          returnedGroupRows: data.returnedrows,
-          totalGroupRows: data.totalrows,
-          groups: data.rows
-        }
-      );
-    })
-    .catch((error) => {
-      console.log(error);
-    })
-  }
-
-  componentWillMount () {
-    this.handleLoadUsers(null, null, null);
-    this.handleLoadGroups(null, null, null);
-  }
-
-  handleUserRowClick(index, name) {
-    if (index === this.state.currentUserIndex) {
-      this.setState({currentUserName: ""});
-      this.setState({currentUserIndex: -1});
-    } else {
-      this.setState({currentUserName: name});
-      this.setState({currentUserIndex: index});
-    }
-  }
-
-  handleGroupRowClick(index, name) {
-    if (index === this.state.currentGroupIndex) {
-      this.setState({currentGroupName: ""});
-      this.setState({currentGroupIndex: -1});
-    } else {
-      this.setState({currentGroupName: name});
-      this.setState({currentGroupIndex: index});
-    }
-  }
-
-  handleMobileUserRowClick (index,name) {
-    this.setState({currentUserName: name});
-    this.setState({currentUserIndex: index});
-    this.setState({deployMobileUserDialog: true});
-  }
-
-  handleMobileGroupRowClick (index, name) {
-    this.setState({currentGroupName: name});
-    this.setState({currentGroupIndex: index});
-    this.setState({deployMobileGroupDialog: true});
   }
 
   render() {
@@ -213,404 +31,11 @@ class Userboard extends React.Component {
 
     return (
       <div>
-        
-        {this.state.deployCreateUser && <CreateUserDialogue handleClose={() => {this.setState({deployCreateUser: false});}}/>}
-        {this.state.deployDeleteUser && <DeleteUserDialogue handleClose={() => {this.setState({deployDeleteUser: false});}} name={this.state.currentUserName}/>}
-        {this.state.deployChangeUserPassword && <ChangeUserPasswordDialogue handleClose={() => {this.setState({deployChangeUserPassword: false});}} name={this.state.currentUserName}/>}
-        {this.state.deployCreateGroup && <CreateGroupDialogue handleClose={() => {this.setState({deployCreateGroup: false});}}/>}
-        {this.state.deployDeleteGroup && <DeleteGroupDialogue handleClose={() => {this.setState({deployDeleteGroup: false});}} name={this.state.currentGroupName}/>}
-        {this.state.deployAddGroupUsers && <AddUserToGroupDialogue handleClose={() => {this.setState({deployAddGroupUsers: false});}} name={this.state.currentGroupName}/>}
-        {this.state.deployRemoveGroupUsers && <RemoveUserFromGroupDialogue handleClose={() => {this.setState({deployRemoveGroupUsers: false});}} name={this.state.currentGroupName}/>}
-        
-        <Hidden mdUp implementation="css">
-          <Dialog
-            open={this.state.deployMobileUserDialog}
-            onClose={() => {this.setState({deployMobileUserDialog: false});}}
-          >
-            <Card>
-              <CardHeader color = "success">
-                {
-                  this.state.currentUserIndex >= 0 && <h2 className={classes.cardTitleWhite}>{this.state.users[this.state.currentUserIndex].name}</h2>
-                }
-              </CardHeader>
-              <CardBody>
-              {
-                this.state.currentUserIndex >= 0 &&
-                <div>
-                  <p className={classes.cardCategory}>Principal Name</p>
-                  <h3 className={classes.cardTitle}>{this.state.users[this.state.currentUserIndex].principalName}</h3>
-
-                  <p className={classes.cardCategory}>Path</p>
-                  <h3 className={classes.cardTitle}>{this.state.users[this.state.currentUserIndex].path}</h3>
-
-                  <p className={classes.cardCategory}>Admin Status</p>
-                  <h3 className={classes.cardTitle}>{this.state.users[this.state.currentUserIndex].isAdmin ? "True" : "False"}</h3>
-
-                  <p className={classes.cardCategory}>Disabled</p>
-                  <h3 className={classes.cardTitle}>{this.state.users[this.state.currentUserIndex].isDisabled ? "True" : "False"}</h3>
-
-                  <Table>
-                    <TableHead>
-                      <TableRow>
-                        <TableCell>Groups</TableCell>
-                      </TableRow>
-                    </TableHead>
-                    <TableBody>
-                      {this.state.users[this.state.currentUserIndex].memberOf.map(
-                        (row) => (
-                          <TableRow
-                            key = {row.name}
-                          >
-                            <TableCell>{row.name}</TableCell>
-                          </TableRow>
-                        )
-                      )}
-                    </TableBody>
-                  </Table>
-                </div>
-              }
-                <GridContainer>
-                  <Button onClick={() => {this.setState({deployDeleteUser: true});}} disabled={this.state.currentUserIndex < 0 ? true:false}>Delete User</Button>
-                  <Button onClick={() => {this.setState({deployChangeUserPassword: true});}} disabled={this.state.currentUserIndex < 0 ? true:false}>Change Password</Button>
-                  <Button onClick={() => {this.setState({deployMobileUserDialog: false});}}>Close</Button>
-                </GridContainer>
-              </CardBody>
-            </Card>
-          </Dialog>
-
-          <Dialog
-            open={this.state.deployMobileGroupDialog}
-            onClose={() =>{this.setState({deployMobileGroupDialog: false});}}
-          >
-            <Card>
-              <CardHeader color="success">
-                {
-                  this.state.currentGroupIndex >= 0 && <h2 className={classes.cardTitleWhite}>{this.state.groups[this.state.currentGroupIndex].name}</h2>
-                }
-              </CardHeader>
-              <CardBody>
-                {
-                  this.state.currentGroupIndex >= 0 &&
-                  <div>
-                    <p className={classes.cardCategory}>Principal Name</p>
-                    <h3 className={classes.cardTitle}> {this.state.groups[this.state.currentGroupIndex].principalName}</h3>
-
-                    <p className={classes.cardCategory}>Path</p>
-                    <h3 className={classes.cardTitle}>{this.state.groups[this.state.currentGroupIndex].path}</h3>
-
-                    <p className={classes.cardCategory}>Members</p>
-                    <h3 className={classes.cardTitle}>{this.state.groups[this.state.currentGroupIndex].members}</h3>
-
-                    <p className={classes.cardCategory}>Declared Members</p>
-                    <h3 className={classes.cardTitle}>{this.state.groups[this.state.currentGroupIndex].declaredMembers}</h3>
-                  </div>
-                }
-                <GridContainer>
-                  <Button onClick={() => {this.setState({deployDeleteGroup: true});}} disabled={this.state.currentGroupIndex < 0 ? true:false}>Delete Group</Button>
-                  <Button onClick={() => {this.setState({deployAddGroupUsers: true});}} disabled={this.state.currentGroupIndex < 0 ? true:false}>Add User to Group</Button>
-                  <Button onClick={() => {this.setState({deployRemoveGroupUsers: true});}} disabled={this.state.currentGroupIndex < 0 ? true:false}>Remove User from Group</Button>
-                  <Button onClick={() => {this.setState({deployMobileGroupDialog: false});}}>Close</Button>
-                </GridContainer>
-              </CardBody>
-            </Card>
-          </Dialog>
-        </Hidden>
-
-
-        <GridContainer>
-          <GridItem xs={12} sm={12} md={5}>
-            <Card>
-              <CardHeader color="warning">
-                <h4 className={classes.cardTitleWhite}>Users</h4>
-              </CardHeader>
-              <CardBody>
-                <Button onClick={() => {this.setState({deployCreateUser: true});}}>Create New User</Button>
-                <form
-                  onSubmit={(event) => { event.preventDefault(); this.handleLoadUsers(this.state.userFilter, null, null);}}
-                >
-                  <TextField
-                    id="user-filter"
-                    name="user-filter"
-                    label="Search User"
-                    onChange={(event) => {this.setState({userFilter: event.target.value});}}
-                  />
-                </form>
-                <Hidden smDown implementation="css">
-                  <Table>
-                    <TableHead>
-                      <TableRow>
-                        {/*this.userColumnNames.map(
-                          row => (
-                            <TableCell
-                              key = {row.id}
-                            >
-                              {row.label}
-                            </TableCell>
-                          )
-                          )*/}
-                        <TableCell>User Names</TableCell>
-                      </TableRow>
-                    </TableHead>
-                    <TableBody>
-                      {this.state.users.map(
-                        (row, index) => (
-                          <TableRow
-                            onClick={(event) => {this.handleUserRowClick(index, row.name);}}
-                            aria-checked={index === this.state.currentUserIndex ? true:false}
-                            key = {row.name}
-                            selected={index === this.state.currentUserIndex ? true:false}
-                          >
-                            <TableCell>
-                                <Checkbox
-                                  checked = {index === this.state.currentUserIndex ? true:false}
-                                />
-                              {row.name}
-                            </TableCell>
-                          </TableRow>
-                        )
-                      )}
-                    </TableBody>
-                    {/*
-                    <TableFooter>
-                      <TableRow>
-                        <TablePagination
-                          rowsPerPageOptions={[3, 10]}
-                          colSpan={1}
-                          count={this.state.totalUserRows}
-                          page=
-                          onChangePage={}
-                        />
-                      </TableRow>
-                    </TableFooter>*/}
-                  </Table>
-                </Hidden>
-                <Hidden mdUp implementation="css">
-                  <Table>
-                    <TableHead>
-                      <TableRow>
-                        {/*this.userColumnNames.map(
-                          row => (
-                            <TableCell
-                              key = {row.id}
-                            >
-                              {row.label}
-                            </TableCell>
-                          )
-                          )*/}
-                        <TableCell>User Names</TableCell>
-                      </TableRow>
-                    </TableHead>
-                    <TableBody>
-                      {this.state.users.map(
-                        (row, index) => (
-                          <TableRow
-                            onClick={(event) => {this.handleMobileUserRowClick(index, row.name);}}
-                            aria-checked={index === this.state.currentUserIndex ? true:false}
-                            key = {row.name}
-                            selected={index === this.state.currentUserIndex ? true:false}
-                          >
-                            <TableCell>
-                              {row.name}
-                            </TableCell>
-                          </TableRow>
-                        )
-                      )}
-                    </TableBody>
-                  </Table>
-                </Hidden>
-              </CardBody>
-            </Card>
-          </GridItem>
-          <GridItem xs={12} sm={12} md={7}>
-            <Hidden smDown implementation="css">
-              <Card>
-                <CardHeader color = "success">
-                  {
-                    this.state.currentUserIndex < 0 ?
-                    <h2 className={classes.cardTitleWhite}>No users selected.</h2>
-                    :
-                    <h2 className={classes.cardTitleWhite}>{this.state.users[this.state.currentUserIndex].name}</h2>
-                  }
-                </CardHeader>
-                <CardBody>
-                {
-                  this.state.currentUserIndex >= 0 &&
-                  <div>
-                    <p className={classes.cardCategory}>Principal Name</p>
-                    <h3 className={classes.cardTitle}>{this.state.users[this.state.currentUserIndex].principalName}</h3>
-
-                    <p className={classes.cardCategory}>Path</p>
-                    <h3 className={classes.cardTitle}>{this.state.users[this.state.currentUserIndex].path}</h3>
-
-                    <p className={classes.cardCategory}>Admin Status</p>
-                    <h3 className={classes.cardTitle}>{this.state.users[this.state.currentUserIndex].isAdmin ? "True" : "False"}</h3>
- 
-                    <p className={classes.cardCategory}>Disabled</p>
-                    <h3 className={classes.cardTitle}>{this.state.users[this.state.currentUserIndex].isDisabled ? "True" : "False"}</h3>
-
-                    <Table>
-                      <TableHead>
-                        <TableRow>
-                          <TableCell>Groups</TableCell>
-                        </TableRow>
-                      </TableHead>
-                      <TableBody>
-                        {this.state.users[this.state.currentUserIndex].memberOf.map(
-                          (row) => (
-                            <TableRow
-                              key = {row.name}
-                            >
-                              <TableCell>{row.name}</TableCell>
-                            </TableRow>
-                          )
-                        )}
-                      </TableBody>
-                    </Table>
-                  </div>
-                }
-                  <GridContainer>
-                    <Button onClick={() => {this.setState({deployDeleteUser: true});}} disabled={this.state.currentUserIndex < 0 ? true:false}>Delete User</Button>
-                    <Button onClick={() => {this.setState({deployChangeUserPassword: true});}} disabled={this.state.currentUserIndex < 0 ? true:false}>Change Password</Button>
-                  </GridContainer>
-                </CardBody>
-              </Card>
-            </Hidden>
-          </GridItem>
-        </GridContainer>
-
-        <GridContainer>
-          <GridItem xs={12} sm={12} md={5}>
-            <Card>
-              <CardHeader color="warning">
-                <h4 className={classes.cardTitleWhite}>Groups</h4>
-              </CardHeader>
-              <CardBody>
-                <Button onClick={() => {this.setState({deployCreateGroup: true});}}>Create New Group</Button>
-                <form
-                  onSubmit={(event) => { event.preventDefault(); this.handleLoadGroups(this.state.groupFilter, null, null);}}
-                >
-                  <TextField
-                    id="group-filter"
-                    name="group-filter"
-                    label="Search Group"
-                    onChange={(event) => {this.setState({groupFilter: event.target.value});}}
-                  />
-                </form>
-                <Hidden smDown implementation="css">
-                  <Table>
-                    <TableHead>
-                      <TableRow>
-                        {/*this.groupColumnNames.map(
-                          row => (
-                            <TableCell
-                              key = {row.id}
-                            >
-                              {row.label}
-                            </TableCell>
-                          )
-                          )*/}
-                        <TableCell>Group Names</TableCell>
-                      </TableRow>
-                      </TableHead>
-                        <TableBody>
-                          {this.state.groups.map(
-                            (row, index) => (
-                              <TableRow
-                                onClick={(event) => {this.handleGroupRowClick(index, row.name);}}
-                                aria-checked={index === this.state.currentGroupIndex ? true : false}
-                                key = {row.name}
-                                selected={index === this.state.currentGroupIndex ? true : false}
-                              >
-                                <TableCell>
-                                    <Checkbox
-                                      checked = {index === this.state.currentGroupIndex ? true : false}
-                                    />
-                                  {row.name}
-                                </TableCell>
-                              </TableRow>
-                            )
-                          )}
-                        </TableBody>
-                  </Table>
-                </Hidden>
-                <Hidden mdUp implementation="css">
-                  <Table>
-                    <TableHead>
-                      <TableRow>
-                        {/*this.groupColumnNames.map(
-                          row => (
-                            <TableCell
-                              key = {row.id}
-                            >
-                              {row.label}
-                            </TableCell>
-                          )
-                          )*/}
-                        <TableCell>Group Names</TableCell>
-                      </TableRow>
-                    </TableHead>
-                    <TableBody>
-                      {this.state.groups.map(
-                        (row, index) => (
-                          <TableRow
-                            onClick={(event) => {this.handleMobileGroupRowClick(index, row.name);}}
-                            aria-checked={index === this.state.currentGroupIndex ? true : false}
-                            key = {row.name}
-                            selected={index === this.state.currentGroupIndex ? true : false}
-                          >
-                            <TableCell>
-                              {row.name}
-                            </TableCell>
-                          </TableRow>
-                        )
-                      )}
-                    </TableBody>
-                  </Table>
-                </Hidden>
-              </CardBody>
-            </Card>
-          </GridItem>
-          <GridItem xs={12} sm={12} md={7}>
-            <Hidden smDown implementation="css">
-              <Card>
-                <CardHeader color="success">
-                  {
-                    this.state.currentGroupIndex < 0 ?
-                    <h2 className={classes.cardTitleWhite}>No group selected.</h2>
-                    :
-                    <h2 className={classes.cardTitleWhite}>{this.state.groups[this.state.currentGroupIndex].name}</h2>
-                  }
-                </CardHeader>
-                <CardBody>
-                  {
-                    this.state.currentGroupIndex >= 0 &&
-                    <div>
-                      <p className={classes.cardCategory}>Principal Name</p>
-                      <h3 className={classes.cardTitle}>{this.state.groups[this.state.currentGroupIndex].principalName}</h3>
-
-                      <p className={classes.cardCategory}>Path</p>
-                      <h3 className={classes.cardTitle}>{this.state.groups[this.state.currentGroupIndex].path}</h3>
-
-                      <p className={classes.cardCategory}>Members</p>
-                      <h3 className={classes.cardTitle}>{this.state.groups[this.state.currentGroupIndex].members}</h3>
-
-                      <p className={classes.cardCategory}>Declared Members</p>
-                      <h3 className={classes.cardTitle}>{this.state.groups[this.state.currentGroupIndex].declaredMembers}</h3>
-                    </div>
-                  }
-                  <GridContainer>
-                    <Button onClick={() => {this.setState({deployDeleteGroup: true});}} disabled={this.state.currentGroupIndex < 0 ? true:false}>Delete Group</Button>
-                    <Button onClick={() => {this.setState({deployAddGroupUsers: true});}} disabled={this.state.currentGroupIndex < 0 ? true:false}>Add User to Group</Button>
-                    <Button onClick={() => {this.setState({deployRemoveGroupUsers: true});}} disabled={this.state.currentGroupIndex < 0 ? true:false}>Remove User from Group</Button>
-                  </GridContainer>
-                </CardBody>
-              </Card>
-            </Hidden>
-          </GridItem>
-        </GridContainer>
+        <UsersManager/>
+        <GroupsManager/>
       </div>
     );
   }
 }
 
-export default withStyles (userboardStyle)(Userboard);
+export default Userboard;
